Feat : 엔티티 클래스 작성 문제, 해설, 질문, 답변, 풀이시도
