+++ src/main/java/Problem/Math/AI/domain/attempt/Status.java
@@ -1,0 +1,17 @@
+package Problem.Math.AI.domain.attempt;
+
+import lombok.Getter;
+
+@Getter
+public enum Status {
+    PENDING("pending"),
+    SUCCESS("success"),
+    FAIL("fail");
+
+    private final String status;
+
+    Status (String status){
+        this.status = status;
+    }
+
+}
